refactor(cli): More refacotring and cleaning
