+++ jaclang/cli/tests/test_cli.py
@@ -2,7 +2,7 @@
 import io
 import sys
 
-from jaclang.cli import cmds
+from jaclang.cli import cli
 from jaclang.utils.test import TestCase
 
 
@@ -19,7 +19,7 @@ class JacCliTests(TestCase):
         sys.stdout = captured_output
 
         # Execute the function
-        cmds.run(self.fixture_abs_path("hello.jac"))  # type: ignore
+        cli.run(self.fixture_abs_path("hello.jac"))  # type: ignore
 
         sys.stdout = sys.__stdout__
         stdout_value = captured_output.getvalue()
@@ -35,14 +35,14 @@ class JacCliTests(TestCase):
 
         # Execute the function
         # try:
-        cmds.enter(self.fixture_abs_path("err2.jac"), entrypoint="speak", args=[])  # type: ignore
+        cli.enter(self.fixture_abs_path("err2.jac"), entrypoint="speak", args=[])  # type: ignore
         # except Exception as e:
         #     print(f"Error: {e}")
 
         sys.stdout = sys.__stdout__
         sys.stderr = sys.__stderr__
         stdout_value = captured_output.getvalue()
-        # print(stdout_value)
+        print(stdout_value)
         # Assertions or verifications
         self.assertIn("Errors occured", stdout_value)
 
@@ -51,11 +51,10 @@ class JacCliTests(TestCase):
         captured_output = io.StringIO()
         sys.stdout = captured_output
 
-        cmds.ast_tool("pass_template")  # type: ignore
+        cli.ast_tool("pass_template")
 
         sys.stdout = sys.__stdout__
         stdout_value = captured_output.getvalue()
-        # print(stdout_value)
         # Assertions or verifications
         self.assertIn("Sub objects.", stdout_value)
         self.assertGreater(stdout_value.count("def exit_"), 10)
