+++ jaclang/cli/tests/test_language.py
@@ -2,7 +2,8 @@
 import io
 import sys
 
-from jaclang.cli import cmds, jac_import
+from jaclang import jac_import
+from jaclang.cli import cli
 from jaclang.utils.test import TestCase
 
 
@@ -19,7 +20,7 @@ class JacLanguageTests(TestCase):
         sys.stdout = captured_output
 
         # Execute the function
-        cmds.run(self.fixture_abs_path("sub_abil_sep.jac"))  # type: ignore
+        cli.run(self.fixture_abs_path("sub_abil_sep.jac"))  # type: ignore
 
         sys.stdout = sys.__stdout__
         stdout_value = captured_output.getvalue()
@@ -36,7 +37,7 @@ class JacLanguageTests(TestCase):
         sys.stdout = captured_output
 
         # Execute the function
-        cmds.run(self.fixture_abs_path("sub_abil_sep_multilev.jac"))  # type: ignore
+        cli.run(self.fixture_abs_path("sub_abil_sep_multilev.jac"))  # type: ignore
 
         sys.stdout = sys.__stdout__
         stdout_value = captured_output.getvalue()
